+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release Notes for Sprig
 
 ## 1.0.0-beta.10 - Unreleased
+### Added
+- Added the ability to override the component trigger using the `trigger` attribute.
+
 ### Changed
 - CSRF tokens are now added as `vars` instead of input fields ([#6](https://github.com/putyourlightson/craft-sprig/issues/6)).
 - The current site is now maintained when rendering components ([#13](https://github.com/putyourlightson/craft-sprig/issues/13)).